+++ planck-cljs/src/planck/repl.cljs
@@ -731,6 +731,12 @@
         cache-json
         sourcemap-json))))
 
+(def ^:private ^:dynamic *source* nil)
+
+(defn- non-shadowing-js-eval
+  []
+  (js/eval *source*))
+
 (defn- js-eval
   [source source-url]
   #_(when (:verbose @app-env)
@@ -739,7 +745,8 @@
     (let [exception (js/PLANCK_EVAL source source-url)]
       (when exception
         (throw exception)))
-    (js/eval source)))
+    (binding [*source* source]
+      (non-shadowing-js-eval))))
 
 (defn- cacheable?
   [{:keys [path name source source-url cache]}]