+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This change
 ## [Unreleased]
 ### Changed
 - transit-cljs 0.8.243
+- Remove `planck.repl/get-arglists` spec (it is non-user supplied and present by default)
 
 ### Fixed
 - Fix issue where source/doc would work on private planck.repl Vars ([#542](https://github.com/mfikes/planck/issues/542))